Implement missing API authorization

+++ pages/api/users/[id]/avatar.ts
@@ -1,6 +1,8 @@
 import Id from '@/models/base/Id'
 import { allowedImageTypes } from '@/models/base/UploadedImage'
 import UserData from '@/models/UserData'
+import MemberPolicy from '@/policies/MemberPolicy'
+import MemberRepo from '@/repos/MemberRepo'
 import UserDataRepo from '@/repos/UserDataRepo'
 import { ApiError } from '@/services/api/ApiErrorService'
 import ApiService, { ApiRequest } from '@/services/ApiService'
@@ -15,16 +17,25 @@ import { v4 as uuid } from 'uuid'
 export default ApiService.handleREST({
   async get(req, res) {
     const id = req.query.id as Id<UserData>
+
     const userData = await UserDataRepo.find(id)
     if (userData === null || userData.avatar === null) {
-      throw new ApiError(404, 'Not Found')
+      return ApiService.Error.notFound()
     }
+
+    const member = await MemberRepo.find(id)
+    if (member === null) {
+      return ApiService.Error.notFound()
+    }
+
+    const policy = ApiService.policy(req, MemberPolicy)
+    ApiService.allowIf(policy.canRead(member))
 
     const fileStream = await AvatarFileService.read(id)
     const fileExt = mime.extension(userData.avatar.mimeType)
     res.writeHead(200, {
       'Content-Type': userData.avatar.mimeType,
-      'Content-Disposition': `inline; filename="avatar.${fileExt}"`,
+      'Content-Disposition': `inline; filename="${member.name}.${fileExt}"`,
     })
     fileStream.pipe(res)
   },
@@ -35,14 +46,23 @@ export default ApiService.handleREST({
       throw new ApiError(415, 'expected \'multipart/form-data\' content')
     }
 
+    const id = req.query.id as Id<UserData>
     const file = await parseFile(req)
 
-    const id = req.query.id as Id<UserData>
     const userData = await UserDataRepo.find(id)
     if (userData === null) {
       await fs.rm(file.filepath)
-      throw new ApiError(404, 'Not Found')
+      return ApiService.Error.notFound()
     }
+
+    const member = await MemberRepo.find(id)
+    if (member === null) {
+      await fs.rm(file.filepath)
+      return ApiService.Error.notFound()
+    }
+
+    const policy = ApiService.policy(req, MemberPolicy)
+    ApiService.allowIf(policy.canEdit(member))
 
     const fileProbe = await probe(createReadStream(file.filepath))
     if (!allowedImageTypes.includes(fileProbe.mime)) {
@@ -79,10 +99,19 @@ export default ApiService.handleREST({
 
   async delete(req, res) {
     const id = req.query.id as Id<UserData>
+
     const userData = await UserDataRepo.find(id)
     if (userData === null || userData.avatar === null) {
-      throw new ApiError(404, 'Not Found')
+      return ApiService.Error.notFound()
     }
+    const member = await MemberRepo.find(id)
+    if (member === null) {
+      return ApiService.Error.notFound()
+    }
+
+    const policy = ApiService.policy(req, MemberPolicy)
+    ApiService.allowIf(policy.canEdit(member))
+
     await AvatarFileService.delete(id)
     await UserDataRepo.update(id, {
       ...userData,
@@ -109,6 +138,7 @@ const parseFile = (req: ApiRequest): Promise<File> => new Promise((resolve) => {
 
 export const config = {
   api: {
+    responseLimit: false,
     bodyParser: false,
   },
 }
